Clientes pt. 2.5

+++ Assets/Resources/Scripts/Managers/ShopManager.cs
@@ -8,14 +8,14 @@ namespace Tcp4.Assets.Resources.Scripts.Managers
     public class ShopManager : Singleton<ShopManager>
     {
         private float stars = 0f;
-        private int money = 0f;
+        private int money = 0;
 
         public void IncreaseMoney(int value) { money += value;}
         public void DecreaseMoney(int value) { money -= value;}
 
         //Getters
         public float GetStars() => stars;
-        public float GetMoney() => money;
+        public int GetMoney() => money;
 
     }
